requirements.txt

scikit-learn
pandas
numpy
matplotlib
seaborn
python-multipart
streamlit
plotly
scipy
pillow